Pass through locked sha again when loading locked kernels
This bit got removed accidentally when adding support for universal
kernels. Also add a test to ensure that we'd catch this in the future.
